+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added: 'cc' debugcommand. Create comments in disassembly with 'cc <line> <comment>'
 - Statusregister can be read/written now by writing to its dataspace address (0x005f)
 
 ## v.0.4.4 - 2020-12-01